+++ components-ui/app/src/tol-ui/src/forms/FormCheckboxes.tsx
@@ -4,7 +4,6 @@ SPDX-FileCopyrightText: 2024 Genome Research Ltd.
 SPDX-License-Identifier: MIT
 */
 
-import { useState } from "react";
 import { Checkbox, CheckboxGroup } from "rsuite";
 import { FormComponentWrapper, TFormCheckboxFields } from "..";
 
@@ -13,7 +12,6 @@ export interface PFormCheckboxes extends TFormCheckboxFields {
   errorText?: string;
   checkedItems: string[];
   setCheckedItems: (value: string[]) => void;
-  defaultChecked?: string[];
 }
 
 export function FormCheckboxes(props: PFormCheckboxes) {
@@ -23,16 +21,13 @@ export function FormCheckboxes(props: PFormCheckboxes) {
     inline,
     indeterminate,
     hidden,
-    defaultChecked = [],
+    checkedItems,
   } = props;
-
-  const [checkedItems, setCheckedItems] = useState(defaultChecked);
 
   const handleCheckboxChange = (value: string) => {
     const updatedCheckedItems = checkedItems.includes(value)
       ? checkedItems.filter((item: string) => item !== value)
       : [...checkedItems, value];
-    setCheckedItems(updatedCheckedItems);
     props.setCheckedItems(updatedCheckedItems);
   };